feat: add validation on to do resource

+++ src/main/java/com/hadiat/livecode5/service/impl/ToDoServiceImpl.java
@@ -8,9 +8,11 @@ import com.hadiat.livecode5.service.AuthenticationService;
 import com.hadiat.livecode5.service.ToDoService;
 import com.hadiat.livecode5.utils.dto.ToDoRequestDTO;
 import com.hadiat.livecode5.utils.dto.ToDoStatusRequestDTO;
+import com.hadiat.livecode5.utils.specifitation.ToDoSpecification;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDate;
@@ -39,23 +41,28 @@ public class ToDoServiceImpl implements ToDoService {
 
     @Override
     public Page<ToDo> findAll(Pageable pageable) {
-        return toDoRepository.findAll(pageable);
-    }
-
-    @Override
-    public Page<ToDo> findAllToDoByAuthUser(Pageable pageable) {
-        return toDoRepository.findAll(pageable);
+        Specification<ToDo> spec = ToDoSpecification.voucherByUser(authenticationService.getUserAuthenticated());
+        return toDoRepository.findAll(spec, pageable);
     }
 
     @Override
     public ToDo findById(Integer id) {
-        return toDoRepository.findById(id).orElseThrow(() -> new RuntimeException("To Do Not Found"));
+        User user = authenticationService.getUserAuthenticated();
+        ToDo toDo = toDoRepository.findById(id).orElseThrow(() -> new RuntimeException("To Do Not Found"));
+        if (!toDo.getUser().equals(user)) {
+            throw new RuntimeException("You don't have agreement to see other todos");
+        }
+        return toDo;
     }
 
 
     @Override
     public ToDo updateById(Integer id, ToDoRequestDTO updatedToDo) {
-        ToDo selectedToDo = findById(id);
+        User user = authenticationService.getUserAuthenticated();
+        ToDo selectedToDo = toDoRepository.findById(id).orElseThrow(() -> new RuntimeException("To Do Not Found"));
+        if (!selectedToDo.getUser().equals(user)) {
+            throw new RuntimeException("You don't have agreement to edit other todos");
+        }
         if (!updatedToDo.getTitle().isEmpty()) selectedToDo.setTitle(updatedToDo.getTitle());
         if (!updatedToDo.getDescription().isEmpty()) selectedToDo.setDescription(updatedToDo.getDescription());
         if (updatedToDo.getStatus() != null) selectedToDo.setStatus(updatedToDo.getStatus());
@@ -78,6 +85,11 @@ public class ToDoServiceImpl implements ToDoService {
 
     @Override
     public void deleteById(Integer id) {
+        User user = authenticationService.getUserAuthenticated();
+        ToDo toDo = toDoRepository.findById(id).orElseThrow(() -> new RuntimeException("To Do Not Found"));
+        if (!toDo.getUser().equals(user)) {
+            throw new RuntimeException("You don't have agreement to see other todos");
+        }
         toDoRepository.deleteById(id);
     }
 }
